+++ updater/tests/nonbot-commits.ps1
@@ -5,7 +5,7 @@ Set-StrictMode -Version latest
 function NonBotCommits([Parameter(Mandatory = $true)][string] $branch)
 {
     $result = & "$PSScriptRoot/../scripts/nonbot-commits.ps1" `
-        -RepoUrl 'https://github.com/getsentry/github-workflows' -MainBranch "main" -PrBranch $branch
+        -RepoUrl 'https://github.com/getsentry/github-workflows' -MainBranch 'main' -PrBranch $branch
     if (-not $?)
     {
         throw $result
@@ -17,17 +17,17 @@ function NonBotCommits([Parameter(Mandatory = $true)][string] $branch)
     $result
 }
 
-RunTest "empty-if-all-commits-by-bot" {
+RunTest 'empty-if-all-commits-by-bot' {
     $commits = NonBotCommits 'deps/updater/tests/sentry-cli.properties'
-    AssertEqual "$commits" ""
+    AssertEqual '' "$commits"
 }
 
-RunTest "empty-if-branch-doesnt-exist" {
+RunTest 'empty-if-branch-doesnt-exist' {
     $commits = NonBotCommits 'non-existent-branch'
-    AssertEqual "$commits" ""
+    AssertEqual '' "$commits"
 }
 
-RunTest "non-empty-if-changed" {
+RunTest 'non-empty-if-changed' {
     $commits = NonBotCommits 'test/nonbot-commits'
-    AssertEqual "$commits" "6133a25 Update README.md"
+    AssertEqual '0b7d9cc test: keep this branch' "$commits"
 }